+++ fixtures/gists-app/app/global-data.ts
@@ -1,5 +1,0 @@
-export function loader({ request }) {
-  return {
-    enableScripts: new URL(request.url).searchParams.get("disableJs") == null
-  };
-}