+++ fixtures/vercel/app/App.js
@@ -1,8 +1,8 @@
-import { Meta, Scripts, Styles, Routes, useGlobalData } from "@remix-run/react";
+import { Meta, Scripts, Styles, Routes, useRouteData } from "@remix-run/react";
 import { Link } from "react-router-dom";
 
 export default function App() {
-  let data = useGlobalData();
+  let data = useRouteData();
 
   return (
     <html lang="en">