+++ packages/core/config.ts
@@ -157,7 +157,11 @@ export async function readConfig(
   let routes = defineConventionalRoutes(appDirectory);
   if (appConfig.routes) {
     let manualRoutes = await appConfig.routes(defineRoutes);
-    routes.push(...manualRoutes);
+    for (let shallowRoute of manualRoutes) {
+      shallowRoute.parentId = "layout:root";
+    }
+    let root = routes[0];
+    (root.children || (root.children = [])).push(...manualRoutes);
   }
 
   let routeManifest = createRouteManifest(routes);