+++ packages/core/serverHelpers.ts
@@ -1,21 +1,12 @@
 import type { AssetManifest } from "./buildManifest";
 import { loadAssetManifest, loadServerManifest } from "./buildManifest";
-import type {
-  ServerEntryModule,
-  GlobalDataModule,
-  RouteModules
-} from "./buildModules";
-import {
-  loadServerEntryModule,
-  loadGlobalDataModule,
-  loadRouteModule
-} from "./buildModules";
+import type { ServerEntryModule, RouteModules } from "./buildModules";
+import { loadServerEntryModule, loadRouteModule } from "./buildModules";
 import type { RemixConfig } from "./config";
 
 export interface ServerBuild {
   assetManifest: AssetManifest;
   serverEntryModule: ServerEntryModule;
-  globalDataModule: GlobalDataModule | null;
   routeModules: RouteModules;
 }
 
@@ -29,7 +20,6 @@ export async function loadServerBuild(
   return {
     assetManifest: loadAssetManifest(buildDir),
     serverEntryModule: loadServerEntryModule(buildDir, manifest),
-    globalDataModule: loadGlobalDataModule(buildDir, manifest),
     routeModules: routeIds.reduce((memo, routeId) => {
       memo[routeId] = loadRouteModule(buildDir, manifest, routeId);
       return memo;