+++ packages/react/browser.tsx
@@ -1,9 +1,8 @@
-import type { ReactElement, ReactNode } from "react";
+import type { ReactElement } from "react";
 import React from "react";
 import type { BrowserHistory, Update } from "history";
 import { createBrowserHistory } from "history";
 import type { EntryContext } from "@remix-run/core";
-import type { ErrorBoundaryComponent } from "@remix-run/core";
 
 import { RemixEntry } from "./components";
 import type { RouteModules } from "./routeModules";
@@ -13,24 +12,12 @@ declare global {
   var __remixRouteModules: RouteModules;
 }
 
-// if ("scrollRestoration" in window.history) {
-//   window.history.scrollRestoration = "manual";
-// }
-
-export interface RemixBrowserProps {
-  children: ReactNode;
-  ErrorBoundary?: ErrorBoundaryComponent;
-}
-
 /**
  * The entry point for a Remix app when it is rendered in the browser (in
  * `entry-browser.js`). This component is used by React to hydrate the HTML
  * that was received from the server.
  */
-export default function RemixBrowser({
-  children,
-  ErrorBoundary
-}: RemixBrowserProps): ReactElement {
+export default function RemixBrowser(): ReactElement {
   let historyRef = React.useRef<BrowserHistory>();
   if (historyRef.current == null) {
     historyRef.current = createBrowserHistory({ window });
@@ -56,12 +43,10 @@ export default function RemixBrowser({
 
   return (
     <RemixEntry
-      children={children}
       context={entryContext}
       action={state.action}
       location={state.location}
       navigator={history}
-      ErrorBoundary={ErrorBoundary}
     />
   );
 }